+++ src/engine/collision/bounding-box.ts
@@ -446,17 +446,7 @@ export class BoundingBox {
   /**
    * Draw a debug bounding box
    * @param ex
-   * @param options
-   * @deprecated
-   */
-  public draw(ex: ExcaliburGraphicsContext, options: RectGraphicsOptions = { color: Color.Yellow }) {
-    ex.debug.drawRect(this.left, this.top, this.width, this.height, options);
-  }
-
-  /**
-   * Draw a debug bounding box
-   * @param ex
-   * @param options
+   * @param [options={ color: Color.Yellow }]
    */
   public debug(ex: ExcaliburGraphicsContext, options: RectGraphicsOptions = { color: Color.Yellow }) {
     ex.debug.drawRect(this.left, this.top, this.width, this.height, options);